some bug fixes and reconfiguration of files
