+++ app/src/main/AndroidManifest.xml
@@ -18,18 +18,14 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".UpdateProfileActivity"
+            android:name=".users.UpdateProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".RegisterActivity"
+            android:name=".users.RegisterActivity"
             android:exported="true">
         </activity>
         <activity
-            android:name=".LoginActivity"
-            android:exported="true">
-        </activity>
-        <activity
-            android:name=".MainActivity"
+            android:name=".users.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -56,16 +52,16 @@
         <activity android:name=".EmailActivity" android:exported="true"/>
 
         <!-- Register AddEventActivity (for adding events) -->
-        <activity android:name=".AddEventActivity" android:exported="true"/>
+        <activity android:name=".organizer.AddEventActivity" android:exported="true"/>
 
         <!-- Register AddFacilityActivity (for adding facilities) -->
-        <activity android:name=".AddFacilityActivity" android:exported="true"/>
+        <activity android:name=".organizer.AddFacilityActivity" android:exported="true"/>
 
         <activity android:name=".FacilityActivity" android:exported="true"/>
 
-        <activity android:name=".EditFacilityActivity" />
+        <activity android:name=".organizer.EditFacilityActivity" />
 
-        <activity android:name=".EventEditActivity" />
+        <activity android:name=".organizer.EventEditActivity" />
