+++ app/src/main/java/com/example/myapplication/FacilityAdapter.java
@@ -7,6 +7,10 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.myapplication.objects.Facility;
+import com.example.myapplication.organizer.EditFacilityActivity;
+
 import java.util.List;
 
 // Adapter class for managing and displaying a list of facilities in a RecyclerView
