--- app/src/main/java/com/example/myapplication/WaitingList.java
@@ -1,4 +1,0 @@
-package com.example.myapplication;
-
-public class WaitingList {
-}
