+++ app/src/main/java/com/example/myapplication/objects/Facility.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.objects;
 
 public class Facility {
     private String id;
