+++ app/src/main/java/com/example/myapplication/objects/Lottery.java
@@ -1,0 +1,4 @@
+package com.example.myapplication.objects;
+
+public class Lottery {
+}
