+++ app/src/main/java/com/example/myapplication/objects/Users.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.objects;
 
 public class Users {
     private String firstName;
