+++ app/src/main/java/com/example/myapplication/objects/WaitingList.java
@@ -1,0 +1,4 @@
+package com.example.myapplication.objects;
+
+public class WaitingList {
+}
