+++ app/src/main/java/com/example/myapplication/organizer/AddFacilityActivity.java
@@ -1,17 +1,18 @@
-package com.example.myapplication;
+package com.example.myapplication.organizer;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-import androidx.annotation.NonNull;
+
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.myapplication.objects.Facility;
+import com.example.myapplication.FacilityActivity;
+import com.example.myapplication.R;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 
 public class AddFacilityActivity extends AppCompatActivity {
