+++ app/src/main/java/com/example/myapplication/organizer/EditFacilityActivity.java
@@ -1,12 +1,14 @@
-package com.example.myapplication;
+package com.example.myapplication.organizer;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.myapplication.objects.Facility;
+import com.example.myapplication.R;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.SetOptions;
