+++ app/src/main/java/com/example/myapplication/users/RegisterActivity.java
@@ -1,7 +1,6 @@
-package com.example.myapplication;
+package com.example.myapplication.users;
 
 import android.content.Intent;
-import android.credentials.CreateCredentialRequest;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.Settings;
@@ -18,19 +17,13 @@ import androidx.activity.result.ActivityResult;
 import androidx.activity.result.ActivityResultCallback;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.myapplication.R;
 import com.example.myapplication.administrator.AdministratorMainActivity;
 import com.example.myapplication.entrant.EntrantMainActivity;
 import com.example.myapplication.organizer.OrganizerMainActivity;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseAuthUserCollisionException;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.storage.FirebaseStorage;
