--- app/src/main/res/layout/activity_login.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/loginLayout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".LoginActivity"
-    android:orientation="vertical"
-    android:gravity="center">
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="30dp"
-        app:cardCornerRadius="30dp"
-        app:cardElevation="20dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:layout_gravity="center_horizontal"
-            android:padding="24dp"
-            android:background="@drawable/login_signup_custom_edittext">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Login"
-                android:textSize="36sp"
-                android:textAlignment="center"
-                android:textColor="@color/lavender"
-                android:textStyle="bold"/>
-
-
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="50dp"
-                android:id="@+id/loginEmail"
-                android:background="@drawable/login_signup_custom_edittext"
-                android:layout_marginTop="40dp"
-                android:padding="8dp"
-                android:hint="Email"
-                android:textColor="@color/black"
-                android:paddingStart="20dp">
-
-            </EditText>
-
-            <EditText
-                android:layout_width="match_parent"
-                android:layout_height="50dp"
-                android:id="@+id/loginPassword"
-                android:background="@drawable/login_signup_custom_edittext"
-                android:padding="8dp"
-                android:paddingStart="20dp"
-                android:layout_marginTop="20dp"
-                android:hint="Password"
-                android:inputType="textPassword"
-                android:textColor="@color/black">
-
-            </EditText>
-
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="50dp"
-                android:id="@+id/loginBtn"
-                android:text="Login"
-                android:padding="8dp"
-                android:layout_marginTop="20dp"
-                android:backgroundTint="@color/lavender"/>
-
-        </LinearLayout>
-    </androidx.cardview.widget.CardView>
-
-    <Button
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:id="@+id/loginSignupBtn"
-        android:text="Sign Up"
-        android:padding="8dp"
-        android:layout_marginHorizontal="30dp"/>
-
-</LinearLayout>
