+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:id="@+id/registerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterActivity"
+    tools:context=".users.RegisterActivity"
     android:orientation="vertical"
     android:gravity="center">
