+++ app/src/main/res/layout/activity_update_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/updateProfileLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UpdateProfileActivity"
+    tools:context=".users.UpdateProfileActivity"
     android:orientation="vertical"
     android:gravity="center">
